Fix concurrency errors that could cause corruption or dropped updates (#1092)
* Explicitly set transaction isolation level to SERIALIZABLE and set eclipselink.transaction.join-existing
Set the Isolation level in both a new EclipseLink SessionCustomizer as well as providing
default postgresql.conf and injecting it in the getting-started docker-compose.

Set eclipselink.transaction.join-existing in persistence.xml to force reads to go through
the same write connection per EntityManager session, otherwise reads are not consistent with
writes in a transaction.

Only call rollback() if the transaction is still active, otherwise we get 500 server errors.

* Fix handling of non-SUCCESS ReturnStatus in BasePolarisCatalog for createTableLike/updateTableLike

Without this fix, even if the underlying database layer does the right thing in resolving
concurrency conflicts, the API call will incorrectly return a 200 success.

This isn't too common of a race condition under normal operation since BasePolarisCatalog
does redundant state-checking right before the transaction (which maybe should be removed
in the future for performance), but in theory this bug indeed could've caused dropped writes
under high concurrency.

* Instead of including the entire sample postgres.conf file, just include
a note that the sample conf file can be found at the postgres github
mirror, and then only specify the settings we immediately care about.

* Add a link to the EclipseLink tutorial for injecting a SessionCustomizer for
attribution of the PolarisEclipseLinkSessionCustomizer.java code.

* Fix format

* Add some other required default postgres.conf settings so that it can start successfully.